+++ NEWS.md
@@ -1,4 +1,4 @@
-# tab2xml (development version)
+# tab2xml 1.1.0
 
 * Add `check_tab()` function.
 * Add `optimize` parameter to `sheet2xml()` function.